whale-tracker/requirements.txt
aiohttp>=3.9.0
requests>=2.31.0
websockets>=12.0
solana>=0.34.0
solders>=0.21.0
python-dotenv>=1.0.0
